Splitting test according to namespaces
